Update page navigation from old domain to new
