+++ src/components/post-item.tsx
@@ -13,7 +13,7 @@ export default function PostItem({ slug, title, publishedAt, summary, descriptio
     <article className="max-w-2xl mx-6 border-b">
       <div>
         <h2 className="text-2xl font-bold">
-          <Link href={slug}>{title}</Link>
+          <Link href={`/ayudantias/${slug}`}>{title}</Link>
         </h2>
       </div>
       <div className="font-regular text-md">{description}</div>
