+++ src/content/ingsw-05.mdx
